[MainPage] 정렬필터 수정 및 리팩토링
