+++ src/components/ComponentBox.jsx
@@ -1,0 +1,140 @@
+import { useState } from "react";
+import {
+  Dialog,
+  DialogBackdrop,
+  DialogPanel,
+  Disclosure,
+  DisclosureButton,
+  DisclosurePanel,
+  Menu,
+  MenuButton,
+  MenuItem,
+  MenuItems,
+} from "@headlessui/react";
+import { XMarkIcon } from "@heroicons/react/24/outline";
+import {
+  ChevronDownIcon,
+  FunnelIcon,
+  MinusIcon,
+  PlusIcon,
+  Squares2X2Icon,
+} from "@heroicons/react/20/solid";
+
+const subCategories = [
+  { name: "Totes", href: "#" },
+  { name: "Backpacks", href: "#" },
+  { name: "Travel Bags", href: "#" },
+  { name: "Hip Bags", href: "#" },
+  { name: "Laptop Sleeves", href: "#" },
+];
+const filters = [
+  {
+    id: "color",
+    name: "Color",
+    options: [
+      { value: "white", label: "White", checked: false },
+      { value: "beige", label: "Beige", checked: false },
+      { value: "blue", label: "Blue", checked: true },
+      { value: "brown", label: "Brown", checked: false },
+      { value: "green", label: "Green", checked: false },
+      { value: "purple", label: "Purple", checked: false },
+    ],
+  },
+  {
+    id: "category",
+    name: "Category",
+    options: [
+      { value: "new-arrivals", label: "샌들 & 슬리퍼", checked: false },
+      { value: "sale", label: "러닝", checked: false },
+      { value: "travel", label: "축구", checked: true },
+      { value: "organization", label: "농구", checked: false },
+      { value: "accessories", label: "트레이닝 & 짐", checked: false },
+      { value: "accessories", label: "골프", checked: false },
+      { value: "accessories", label: "테니스", checked: false },
+    ],
+  },
+  {
+    id: "size",
+    name: "Size",
+    options: [
+      { value: "2l", label: "2L", checked: false },
+      { value: "6l", label: "6L", checked: false },
+      { value: "12l", label: "12L", checked: false },
+      { value: "18l", label: "18L", checked: false },
+      { value: "20l", label: "20L", checked: false },
+      { value: "40l", label: "40L", checked: true },
+    ],
+  },
+];
+
+<section aria-labelledby="products-heading" className="pb-24 pt-6">
+  <h2 id="products-heading" className="sr-only">
+    Products
+  </h2>
+
+  <div className="grid grid-cols-1 gap-x-8 gap-y-10 lg:grid-cols-4">
+    {/* Filters */}
+    <form className="hidden lg:block">
+      <h3 className="sr-only">Categories</h3>
+      <ul
+        role="list"
+        className="space-y-4 border-b border-gray-200 pb-6 text-sm font-medium text-gray-900"
+      >
+        {subCategories.map((category) => (
+          <li key={category.name}>
+            <a href={category.href}>{category.name}</a>
+          </li>
+        ))}
+      </ul>
+
+      {filters.map((section) => (
+        <Disclosure
+          key={section.id}
+          as="div"
+          className="border-b border-gray-200 py-6"
+        >
+          <h3 className="-my-3 flow-root">
+            <DisclosureButton className="group flex w-full items-center justify-between bg-white py-3 text-sm text-gray-400 hover:text-gray-500">
+              <span className="font-medium text-gray-900">{section.name}</span>
+              <span className="ml-6 flex items-center">
+                <PlusIcon
+                  aria-hidden="true"
+                  className="h-5 w-5 group-data-[open]:hidden"
+                />
+                <MinusIcon
+                  aria-hidden="true"
+                  className="h-5 w-5 [.group:not([data-open])_&]:hidden"
+                />
+              </span>
+            </DisclosureButton>
+          </h3>
+          <DisclosurePanel className="pt-6">
+            <div className="space-y-4">
+              {section.options.map((option, optionIdx) => (
+                <div key={option.value} className="flex items-center">
+                  <input
+                    defaultValue={option.value}
+                    defaultChecked={option.checked}
+                    id={`filter-${section.id}-${optionIdx}`}
+                    name={`${section.id}[]`}
+                    type="checkbox"
+                    className="h-4 w-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-500"
+                  />
+                  <label
+                    htmlFor={`filter-${section.id}-${optionIdx}`}
+                    className="ml-3 text-sm text-gray-600"
+                  >
+                    {option.label}
+                  </label>
+                </div>
+              ))}
+            </div>
+          </DisclosurePanel>
+        </Disclosure>
+      ))}
+    </form>
+
+    {/* Product grid */}
+    <div className="lg:col-span-3">{/* Your content */}</div>
+  </div>
+</section>;
