+++ src/components/MockProduct.jsx
@@ -1,12 +1,4 @@
-// mockData.js
-
-// 원화로 변환하는 함수
-function formatToKRW(amount) {
-  return new Intl.NumberFormat("ko-KR", {
-    style: "currency",
-    currency: "KRW",
-  }).format(amount);
-}
+import formatToKRW from "../utils/FormatToKRW";
 
 // 목업데이터
 export const MockProduct = {
